@@ -15,30 +15,31 @@
 
 name: instant_xml_parser
 description: Code generator for instant_xml_parser. Reads specially annotated Dart code an emits an XML parser/deserializer to match.
-version: 1.0.0+2
+version: 1.1.0
 homepage: https://github.com/sarahec/instant_xml_parser
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: '>=0.39.11 <0.42.0'
-  build: ^1.3.0
-  build_config: ^0.4.0
+  analyzer: ^1.7.2
+  build: ^2.0.3
+  build_config: ^1.0.0
   built_collection: ^5.0.0
-  built_value: ^8.0.0
-  code_builder: ^3.0.0
-  dart_style: ^1.0.0
-  logging: ^0.11.4
+  built_value: ^8.1.0
+  code_builder: ^4.1.0
+  collection: ^1.15.0
+  dart_style: ^2.0.3
+  logging: ^1.0.1
   meta: ^1.2.4
-  recase: ^3.0.0
-  ixp_runtime: '>=0.11.0 <2.0.0'
-  source_gen: ^0.9.6
-  xml: ^4.4.1
+  recase: ^4.0.0
+  ixp_runtime: ^1.1.0
+  source_gen: ^1.0.3
+  xml: ^5.1.2
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  build_test: ^1.2.1
-  built_value_generator: ^8.0.0
+  build_runner: ^2.0.6
+  build_test: ^2.1.2
+  built_value_generator: ^8.1.0
   pedantic: ^1.9.0
   test: ^1.0.0
