@@ -18,7 +18,7 @@ import 'package:instant_xml_parser/src/generators/parse_method_generator.dart';
 
 /// Entry point to the generation process. Called from ```build.yaml```.
 Builder parserBuilder(BuilderOptions options) {
-  final optionsMap = Map<String, dynamic>.from(options?.config ?? {});
+  final optionsMap = Map<String, dynamic>.from(options.config);
   final useNullSafety = optionsMap.remove('use_null_safety') as bool;
   return LibraryBuilder(ParseMethodGenerator(useNullSafety),
       generatedExtension: '.parser.dart');
