@@ -16,12 +16,16 @@ import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart' show AssetId;
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
+import 'package:collection/collection.dart';
+import 'package:logging/logging.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'class_info.dart';
 import 'method_info.dart';
 
 part 'source_info.g.dart';
+
+final _log = Logger('SourceInfo');
 
 /// Per-file information used by the parser. Serialiazable so that the
 /// first-pass builder can output this as JSON then reconstitute
@@ -33,10 +37,15 @@ abstract class SourceInfo implements Built<SourceInfo, SourceInfoBuilder> {
   factory SourceInfo([void Function(SourceInfoBuilder) updates]) = _$SourceInfo;
 
   factory SourceInfo.fromLibrary(LibraryReader reader, AssetId asset) {
-    return SourceInfo((b) => b
-      ..element = reader.element
-      ..classes = [for (var c in reader.classes) ClassInfo.fromElement(c)]
-      ..uri = asset.uri);
+    final classes = [for (var c in reader.classes) ClassInfo.fromElement(c)];
+    _log.finer(
+        'SourceInfo(element: ${reader.element}, classes: $classes, uri: ${asset.uri})');
+    return SourceInfo((b) {
+      b
+        ..element = reader.element
+        ..classes = classes
+        ..uri = asset.uri;
+    });
   }
 
   SourceInfo._();
@@ -46,19 +55,30 @@ abstract class SourceInfo implements Built<SourceInfo, SourceInfoBuilder> {
   LibraryElement get element;
 
   @memoized
-  Iterable<String> get importUris => element.imports.map((i) => i.uri);
+  Iterable<String> get importUris => [
+        for (var i in element.imports.where((e) => e.uri != null)) i.uri!
+      ]..sort();
 
   @memoized
-  Iterable<MethodInfo> get methods => classes
-      .map((c) => c.method)
-      .where((c) => c != null)
-      .toList(growable: false)
-        ..sort((a, b) => a.name.compareTo(b.name));
+  Iterable<MethodInfo> get methods {
+    final result = [
+      for (var c in classes.where((p) => p.method != null)) c.method!
+    ]..sort((a, b) => a.name.compareTo(b.name));
+    _log.finer('methods -> $result');
+    return result;
+  }
 
   Uri get uri;
 
-  ClassInfo classForType(DartType t) =>
-      classes.firstWhere((v) => v.type == t, orElse: () => null);
+  /// Look up the class info object, ignoring the nullability of `t`
+  ClassInfo? classForType(DartType t) {
+    final t_name = t.getDisplayString(withNullability: false);
+    final result = classes.firstWhereOrNull(
+        (v) => v.type.getDisplayString(withNullability: false) == t_name);
+    _log.finer(() =>
+        'classForType(${t.getDisplayString(withNullability: true)}) -> $result');
+    return result;
+  }
 
   ClassInfo classNamed(name) => classes.firstWhere((c) => c.typeName == name);
 
@@ -66,13 +86,27 @@ abstract class SourceInfo implements Built<SourceInfo, SourceInfoBuilder> {
 
   Iterable<MethodInfo> methodsReturning(DartType desiredType,
       [allowSubtypes = true]) {
-    var result = methods.where((m) => m.type == desiredType);
+    // We have a problem: returned types are non-nullable, but a field's type
+    // may be. So, convert both to type names without nullability and match.
+    // (It may be hacky, but requires the least implementation knowledge of
+    // the type system, and is therefore safest.)
+    final desiredTypeName =
+        desiredType.getDisplayString(withNullability: false);
+    var result = methods.where((m) =>
+        m.type.getDisplayString(withNullability: false) == desiredTypeName);
     if (result.isEmpty && allowSubtypes) {
-      result = subclassesOf(desiredType)?.map((c) => c.method);
+      _log.finest('No methods found on $desiredType, looking for subclasses');
+      result = [
+        for (var c
+            in subclassesOf(desiredType).where((sc) => sc.method != null))
+          c.method!
+      ];
     }
-    return (result == null || result.isEmpty) ? null : result;
+    _log.finer(
+        'methodsReturning($desiredType, allowSubtypes: $allowSubtypes) -> $result');
+    return result;
   }
 
   Iterable<ClassInfo> subclassesOf(DartType type) =>
-      classes.where((ClassInfo c) => c.type.superclass == type).toList();
+      classes.where((ClassInfo c) => c.type.superclass == type);
 }
